Fix like function

+++ app/Http/Controllers/LikeController.php
@@ -16,38 +16,45 @@ class LikeController extends Controller
     public function updateLikeStatus($postId) {
         $user = Session::get('sUser');
         if (isset($user)) {
+            $post = Post::find($postId);
             $likePost = LikePost::where([
                 ['user_id', '=', $user->id],
                 ['post_id', '=', $postId],
             ])->first();
             if (isset($likePost)) {
                 LikePost::destroy($likePost->id);
-                $notification = Notification::where([
-                    ['fuser_id' => $user->id],
-                    ['post_id' => $postId],
-                ])->first();
-                Notification::destroy($notification->id);
+                $notification = Notification::where('post_id', '=', $postId)
+                    ->where('user_id', '=', $post->user_id)
+                    ->where('fuser_id', '=', $user->id)
+                    ->where('content', '=', $user->name . ' liked your status')->first();
+                if (isset($notification)) {
+                    Notification::destroy($notification->id);
+                }
             } else {
                 $likePost = new LikePost([
                     'user_id' => $user->id,
                     'post_id' => $postId,
                 ]);
                 $likePost->save();
-                $post = Post::where('id', $postId)->first();
-                $newNotification = new Notification([
-                    'user_id' => $post->user_id,
-                    'fuser_id' => $user->id,
-                    'post_id' => $post->id,
-                    'content' => $user->name . ' liked your status',
-                ]);
-                $newNotification->save();
+                $notification = Notification::where('post_id', '=', $postId)
+                    ->where('user_id', '=', $post->user_id)
+                    ->where('fuser_id', '=', $user->id)
+                    ->where('content', '=', $user->name . ' liked your status')->first();
+                if(!isset($notification)) {
+                    $newNotification = new Notification([
+                        'user_id' => $post->user_id,
+                        'fuser_id' => $user->id,
+                        'post_id' => $post->id,
+                        'content' => $user->name . ' liked your status',
+                    ]);
+                    $newNotification->save();
+                }
             }
-            $post = Post::find($postId);
             if(!isset($post)){
                 Session::put('message', "Have something wrong!");
                 return redirect('home');
             } else {
-                $likePostNumber = count(LikePost::where('post_id', '=', $post->id)->get());
+                $likePostNumber = count(LikePost::where('post_id', '=', $postId)->get());
                 return response()->json($likePostNumber);
             }
         } else return redirect('home');
@@ -57,16 +64,18 @@ class LikeController extends Controller
     {
         $user = Session::get('sUser');
         if (isset($user)) {
+            $comment = Comment::find($commentId);
             $likeComment = LikeComment::where([
                 ['user_id', '=', $user->id],
                 ['comment_id', '=', $commentId],
             ])->first();
             if (isset($likeComment)) {
                 likeComment::destroy($likeComment->id);
-                $notification = Notification::where([
-                    ['fuser_id' => $user->id],
-                    ['comment_id' => $commentId],
-                ])->first();
+                $notification = Notification::where('comment_id', '=', $commentId)
+                    ->where('post_id', '=', $comment->post_id)
+                    ->where('user_id', '=', $comment->user_id)
+                    ->where('fuser_id', '=', $user->id)
+                    ->where('content', '=', $user->name . ' liked your comment')->first();
                 Notification::destroy($notification->id);
             } else {
                 $likeComment = new likeComment([
@@ -74,15 +83,21 @@ class LikeController extends Controller
                     'comment_id' => $commentId
                 ]);
                 $likeComment->save();
-                $comment = Comment::where('id', $commentId)->first();
-                $newNotification = new Notification([
-                    'user_id' => $comment->user_id,
-                    'fuser_id' => $user->id,
-                    'post_id' => $comment->post_id,
-                    'comment_id' => $comment->id,
-                    'content' => $user->name . ' liked your comment',
-                ]);
-                $newNotification->save();
+                $notification = Notification::where('comment_id', '=', $commentId)
+                    ->where('post_id', '=', $comment->post_id)
+                    ->where('user_id', '=', $comment->user_id)
+                    ->where('fuser_id', '=', $user->id)
+                    ->where('content', '=', $user->name . ' liked your comment')->first();
+                if (!$notification) {
+                    $newNotification = new Notification([
+                        'user_id' => $comment->user_id,
+                        'fuser_id' => $user->id,
+                        'post_id' => $comment->post_id,
+                        'comment_id' => $comment->id,
+                        'content' => $user->name . ' liked your comment',
+                    ]);
+                    $newNotification->save();
+                }
             }
             $comment = Comment::find($commentId);
             if(!isset($comment)){
